Finalizing manager
