+++ .gitignore
@@ -160,6 +160,5 @@ database/online/keys/
 database/experiments
 
 data/db/*.db
-data/json/*.json
 data/zip/*.bz2
